+++ src/useIsHotkeyPressed.ts
@@ -1,0 +1,5 @@
+import hotkeys from "hotkeys-js";
+
+export function useIsHotkeyPressed() {
+  return hotkeys.isPressed;
+}